requirements.txt

black>=20.8b1
fastapi>=0.61
flake8>=3.8
imageio>=2.9
ipython>=7.23
jupyter>=1.0
matplotlib-base>=3.4
networkx>=2.5
notebook>=6.3
notexbook-theme>=2.0
nptyping>=1.3
numpy>=1.19
pandas>=1.1
pandoc>=2.13
pillow>=7.2
pip>=20.2
pytorch>=1.6
pre-commit>=2.7
pydantic>=1.8
pyflakes>=2.2
pytest>=6.1
requests>=2.25
rinohtype>=0.5
scikit-image>=0.17
scikit-learn>=0.23
scipy>=1.5
snowballstemmer>=2.1
setuptools>=49.6
sphinx>=3.2
sqlite>=3.35
starlette>=0.13
toml>=0.10
torchvision>=0.7
uvicorn>=0.12
yaml>=0.2

